ndvi-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
earthengine-api==0.1.392
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
requests==2.31.0
APScheduler==3.10.4
pydantic==2.5.0
